Elysium server-side setup:
This commit sets up all the basic folders and files for the server side code. This includes installing all the necessary modules: 1. express, 2. cors, 3. mongodb, 4. dotenv, 5. jsonwebtoken, 6. cookie-parser. This commit also creates a gitignore file which is set to ignore the node_modules folder and the .env and .vercel files.